Implemented a core set of network processing tools in Python
The module nws_tools contains a core set network visualization/generation
and analysis tools. Already proved to be quite useful. The routine imwrite
in imtools was also modified and updated to better reflect changes in
the default saving behavior of matplotlib. The global init file was
updated to import nws_tools as nwt by default.