Modified and cleaned up docstrings of whole package
In the course of archiving codes in the lab's git repository I seized
the chance to clean up and rewrite the docstrings of the functions in here.
Now everything uses "Inputs" as section header to explain the input
variables (the NumPy-style "Parameters" is a little confusing anyway).
Also http-links are now converted to Sphinx-style syntax to make them
clickable in a sphinxed document.