Modified mean-network computation
The code in get_meannw was overhauled to account for connectedness of the
mean network. Now if the mean network is not connected for a certain
percentage value percval, percval is decreases by 5%, then the computation
is repeated. If the average network is still disconnected for percval=0%
then at least one node is disconnected in all per-subject networks and a
ValueError is raised.