Bugfix in nws_tools.py
As Joel pointed out correctly, Axes3D has to be imported for
projection=3d to work correctly. Since Axes3D is available in any working
matplotlib installation, Axes3D was added as top-level import. Further,
the import from recipes was updated to reflect whether the code is run
locally on my computer (with mypy available) or on the server.