Initial commmit of sim_tools
After much pain, here is the first version of sim_tools, to be used
with the NM model. Since this is not intended to be made public any time
soon, there are no docstrings or any other super fancy bells and whistles
included.
Also, in the course of writing the chunking code for sim_tools, a bug
in eeg_tools was fixed:
* Fixed bug that made blocksize in eeg_tools.py too long
  (if mod(numSamples,maxmem) == 0 then,
  the blocksize list gets a 0 appended, which doesn't make much sense...)